Refactor and add logging option

+++ internal/log/log.go
@@ -1,6 +1,7 @@
 package log
 
 import (
+	"bufio"
 	"fmt"
 	"os"
 	"sync"
@@ -19,24 +20,49 @@ const (
 
 const maxLogSize = 10 << 20 // 10 MB
 
-var (
-	logFile *os.File
-	mu      sync.Mutex
+const (
+	logModeFull    = "full"
+	logModeMinimal = "minimal"
+	logModeOff     = "off"
+
+	logBufferSize  = 4096
+	logFlushPeriod = 250 * time.Millisecond
 )
 
-func SetOutput(path string) error {
+var (
+	logFile    *os.File
+	logMode    = logModeFull
+	logCh      chan string
+	stopWriter chan struct{}
+	writerWG   sync.WaitGroup
+	mu         sync.RWMutex
+)
+
+func SetOutput(path string, mode string) error {
 	mu.Lock()
 	defer mu.Unlock()
 
+	shutdownWriterLocked()
+	logMode = normalizeMode(mode)
+	if logMode == logModeOff {
+		return nil
+	}
+
 	if info, err := os.Stat(path); err == nil && info.Size() > maxLogSize {
-		os.Truncate(path, 0)
+		_ = os.Truncate(path, 0)
 	}
 
 	f, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
 	if err != nil {
 		return err
 	}
+
 	logFile = f
+	logCh = make(chan string, logBufferSize)
+	stopWriter = make(chan struct{})
+
+	writerWG.Add(1)
+	go writerLoop(f, logCh, stopWriter)
 	return nil
 }
 
@@ -44,10 +70,7 @@ func Close() {
 	mu.Lock()
 	defer mu.Unlock()
 
-	if logFile != nil {
-		logFile.Close()
-		logFile = nil
-	}
+	shutdownWriterLocked()
 }
 
 func ColorForStatus(code int) string {
@@ -64,15 +87,28 @@ func ColorForStatus(code int) string {
 }
 
 func Request(domain string, method string, path string, upstream int, status int, duration time.Duration) {
+	mu.RLock()
+	mode := logMode
+	ch := logCh
+	mu.RUnlock()
+
+	if mode == logModeOff || ch == nil {
+		return
+	}
+
 	ts := time.Now().Format("15:04:05")
 	dur := formatDuration(duration)
 
-	mu.Lock()
-	if logFile != nil {
-		fmt.Fprintf(logFile, "%s\t%s\t%s\t%s\t%d\t%d\t%s\n",
-			ts, domain, method, path, upstream, status, dur)
+	line := fmt.Sprintf("%s\t%s\t%s\t%s\t%d\t%d\t%s\n",
+		ts, domain, method, path, upstream, status, dur)
+	if mode == logModeMinimal {
+		line = fmt.Sprintf("%s\t%s\t%d\t%s\n", ts, domain, status, dur)
 	}
-	mu.Unlock()
+
+	select {
+	case ch <- line:
+	default:
+	}
 }
 
 func Info(format string, args ...interface{}) {
@@ -91,4 +127,65 @@ func formatDuration(d time.Duration) string {
 		return fmt.Sprintf("%dms", d.Milliseconds())
 	}
 	return fmt.Sprintf("%.1fs", d.Seconds())
+}
+
+func normalizeMode(mode string) string {
+	switch mode {
+	case logModeFull, "":
+		return logModeFull
+	case logModeMinimal:
+		return logModeMinimal
+	case logModeOff:
+		return logModeOff
+	default:
+		return logModeFull
+	}
+}
+
+func writerLoop(file *os.File, entries <-chan string, stop <-chan struct{}) {
+	defer writerWG.Done()
+
+	buffered := bufio.NewWriterSize(file, 64*1024)
+	ticker := time.NewTicker(logFlushPeriod)
+	defer ticker.Stop()
+
+	flushAndClose := func() {
+		_ = buffered.Flush()
+		_ = file.Close()
+	}
+
+	for {
+		select {
+		case line := <-entries:
+			if _, err := buffered.WriteString(line); err != nil {
+				flushAndClose()
+				return
+			}
+		case <-ticker.C:
+			_ = buffered.Flush()
+		case <-stop:
+			for {
+				select {
+				case line := <-entries:
+					if _, err := buffered.WriteString(line); err != nil {
+						flushAndClose()
+						return
+					}
+				default:
+					flushAndClose()
+					return
+				}
+			}
+		}
+	}
+}
+
+func shutdownWriterLocked() {
+	if stopWriter != nil {
+		close(stopWriter)
+		writerWG.Wait()
+		stopWriter = nil
+	}
+	logCh = nil
+	logFile = nil
 }
